+++ test/appcache/appcache_impl_test.dart
@@ -3,23 +3,23 @@ import 'package:test/test.dart';
 
 void main() {
   group('appcache', () {
-    test('excluded', () {
+    test('exclude', () {
       var map = settingsAddExcluded(null, ['test']);
       expect(map, {
-        'excluded': ['test']
+        'exclude': ['test']
       });
       map = settingsAddExcluded(map, ['test2']);
       expect(map, {
-        'excluded': ['test', 'test2']
+        'exclude': ['test', 'test2']
       });
     });
 
     test('fix_settings', () async {
       var map = await fixAppCacheSettings({
-        'appcache_excluded': ['x']
+        'appcache_exclude': ['x']
       });
       expect(map, {
-        'excluded': ['x']
+        'exclude': ['x']
       });
     });