+++ lib/pub/global.dart
@@ -2,8 +2,10 @@ import 'package:tekartik_build_utils/common_import.dart';
 
 Future<bool> isActivated(String packageName, {bool verbose}) async {
   var lines = LineSplitter.split(
-      (await runCmd(PubCmd(['global', 'list']), verbose: verbose)).stdout
-          as String);
+      (await runCmd(PubCmd(['global', 'list']), verbose: verbose))
+          .stdout
+          .toString()
+          .trim());
   for (var line in lines) {
     if (line.split(' ')[0] == packageName) {
       return true;