+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## 0.6.4
+## 0.6.6
 
+- support `appcache_exclude` in fixAppCache
 - add pub utils to activate a package if needed
 
 ## 0.6.0